requirements.txt
Flask==0.10.1
Jinja2==2.7.3
MarkupSafe==0.23
Pillow==2.8.1
Werkzeug==0.10.4
itsdangerous==0.24
pilkit==1.1.12
selenium==2.45.0
six==1.9.0
wsgiref==0.1.2
